AOJ ITP1_4_B rにto_fを使用(AC) (#41)

+++ AOJ/ITP1/4_B.rb
@@ -1,4 +1,4 @@
-r = gets.to_i
+r = gets.to_f
 
 pi = Math::PI
 s = r * r * pi
